[Core] Rename utils/children.js
Rename utils/children.js to utils/childUtils.js
Convert to named exports, update imports and usage
