+++ src/RaisedButton/RaisedButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import transitions from '../styles/transitions';
 import ColorManipulator from '../utils/colorManipulator';
-import Children from '../utils/children';
+import {createChildFragment} from '../utils/childUtils';
 import EnhancedButton from '../internal/EnhancedButton';
 import Paper from '../Paper';
 import getMuiTheme from '../styles/getMuiTheme';
@@ -395,7 +395,7 @@ const RaisedButton = React.createClass({
         iconCloned,
         labelElement,
       };
-    const enhancedButtonChildren = Children.create(childrenFragment);
+    const enhancedButtonChildren = createChildFragment(childrenFragment);
 
     return (
       <Paper
